pyproject.toml
[project]
name = "customer-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "chainlit>=2.7.2",
    "chromadb>=1.0.20",
    "faker>=37.6.0",
    "fastmcp>=2.11.3",
    "gradio>=5.44.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "sqlite-utils>=3.38",
    "uvicorn>=0.35.0",
]
